+++ packages/at_lookup/lib/at_lookup.dart
@@ -1,9 +1,8 @@
 export 'src/at_lookup.dart';
-export 'src/at_lookup_impl.dart';
+export 'src/cache/cacheable_secondary_address_finder.dart';
+export 'src/cache/secondary_address_finder.dart';
+export 'src/connection/at_lookup_socket_factories.dart';
 export 'src/connection/outbound_connection.dart';
 export 'src/connection/outbound_connection_impl.dart';
 export 'src/exception/at_lookup_exception.dart';
-export 'src/monitor_client.dart';
-export 'src/cache/secondary_address_finder.dart';
-export 'src/cache/cacheable_secondary_address_finder.dart';
 export 'src/util/secure_socket_util.dart';